+++ tools/Azure.Mcp.Tools.AzureManagedLustre/src/Commands/AzureManagedLustreJsonContext.cs
@@ -10,8 +10,10 @@ namespace Azure.Mcp.Tools.AzureManagedLustre.Commands;
 [JsonSerializable(typeof(FileSystemSubnetSizeCommand.FileSystemSubnetSizeResult))]
 [JsonSerializable(typeof(FileSystemListCommand.FileSystemListResult))]
 [JsonSerializable(typeof(SkuGetCommand.SkuGetResult))]
-[JsonSerializable(typeof(LustreFileSystem))]
 [JsonSerializable(typeof(AzureManagedLustreSkuInfo))]
 [JsonSerializable(typeof(AzureManagedLustreSkuCapability))]
+[JsonSerializable(typeof(FileSystemImportJobCreateCommand.FileSystemImportJobCreateResult))]
+[JsonSerializable(typeof(LustreFileSystem))]
+[JsonSerializable(typeof(ImportJobInfo))]
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase, WriteIndented = true, DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull)]
 internal partial class AzureManagedLustreJsonContext : JsonSerializerContext;